+++ packages/shared/src/components/modal/index.ts
@@ -1,0 +1,4 @@
+export { default as AdminModals } from './AdminModals';
+export { default as ClientModals } from './ClientModals';
+export { default as ScoreCalculationModal } from './ScoreCalculationModal';
+export { default as SharedModals } from './SharedModals';
